+++ src/test/java/com/wap/app2/tagogayo/GachitayoApplicationTests.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class TagogayoApplicationTests {
+class GachitayoApplicationTests {
 
 	@Test
 	void contextLoads() {
